Implemented a parentHook for Quiz queries
