+++ client/src/routes/dashboard/pages/quiz/QuizPage.tsx
@@ -1,13 +1,13 @@
 import { useParams } from "react-router-dom";
 import { URLS } from "../../../../configs/Links";
 import LoadingScreen from "../../../../shared/components/LoadingScreen";
-import { useQuizFetch } from "../../../../shared/hooks/api/useQuizFetch.hook";
+import { useQuizQueries } from "../../../../shared/hooks/api/useQuizQueries.hook";
 import RedirectErrorPage from "../../../error/RedirectErrorPage";
 import Quiz from "./Quiz";
 
 export default function QuizPage() {
   const { quizID } = useParams();
-  const { quizData, isPending, isError } = useQuizFetch(quizID!);
+  const { quizData, isPending, isError } = useQuizQueries().useQuizGameData(quizID!);
 
   if (isError)
     return (
